Added Function field to EditorScript
now the Editor script can get the function name from the script
components and adds them to the list. The script then uses a send
message to execute that code. I'll probably switch to a hashtable to
filter the functions by component selected. After that probably work on
optimizing it.